Remove the RrSyncState and RrPendingChange models
This also involves editing migrations which used those
models directly.

+++ db/migrate/20151208150956_fix_host_storage_replication_on_upgrade.rb
@@ -1,6 +1,5 @@
 class FixHostStorageReplicationOnUpgrade < ActiveRecord::Migration
   include MigrationHelper
-  include MigrationHelper::SharedStubs
 
   class MiqRegion < ActiveRecord::Base; end
 
@@ -9,36 +8,10 @@ class FixHostStorageReplicationOnUpgrade < ActiveRecord::Migration
   end
 
   def up
-    if on_replication_target?
-      run_for_replication_target
-    else
-      run_for_replication_source
-    end
+    HostsStorage.delete_all if on_replication_target?
   end
 
   def on_replication_target?
     MiqRegion.select(:region).distinct.count > 1
-  end
-
-  def run_for_replication_target
-    HostsStorage.delete_all
-  end
-
-  def run_for_replication_source
-    return unless RrSyncState.table_exists?
-
-    prefix = "rr#{ApplicationRecord.my_region_number}"
-    old_name = "hosts_storages"
-    new_name = "host_storages"
-
-    drop_trigger(new_name, "#{prefix}_#{old_name}")
-
-    say_with_time("Deleting pending changes for #{old_name}") do
-      RrPendingChange.where(:change_table => old_name).delete_all
-    end
-
-    say_with_time("Deleting sync state for #{old_name}") do
-      RrSyncState.where(:table_name => old_name).delete_all
-    end
   end
 end
